Managed to add some colour to the visvis python rendering.
It's quite bad! It thresholds the volume and makes an RGB colume with
a colour value associated in each voxel. Happy to accept patches if
you know of a better way with visvis.

Also, the marching cubes in visvis doesn't appear to apply any
smoothing to the volume, unlike MATLAB, so the renderings will look
very blocky.